Feature/am 167 delete patient

+++ src/app/patients/specific-patient/specific-patient.page.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
+import { AlertController } from '@ionic/angular';
 import { ResultsApiService } from 'src/app/results/shared-results/services/results-api/results-api.service';
+import { DialogsComponent } from 'src/app/shared/components/dialogs/dialogs.component';
 import { PatientsApiService } from '../shared-patients/services/patients-api/patients-api.service';
 
 export interface Patient {
@@ -37,7 +39,9 @@ export class SpecificPatientPage implements OnInit {
     private patientsApiService: PatientsApiService,
     private resultsApiService: ResultsApiService,
     private route: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    public alertController: AlertController,
+    private dialogsComponent: DialogsComponent
   ) {}
 
   ngOnInit() {}
@@ -63,11 +67,17 @@ export class SpecificPatientPage implements OnInit {
   /**
    * Borra el paciente (cuando clickea el botón de eliminar).
    */
-  deletePatient() {
+  async deletePatient() {
+    const confirm = await this.dialogsComponent.presentAlertConfirm('Paciente',
+    '¿Desea eliminar al paciente? Esta acción no puede deshacerse')
+      
+    if (confirm) {
     this.patientsApiService.deletePatient(this.id).subscribe(() => {
-      this.router.navigateByUrl('/patients');
-    });
+        this.router.navigateByUrl('/patients');
+      });
+    }
   }
+
 
   /**
    * Desvincula el paciente.
@@ -97,5 +107,4 @@ export class SpecificPatientPage implements OnInit {
       .subscribe((res) => {});
   }
 
-  
 }
